Cleaned up cython implementation

+++ python/cython/c_dwt.pxd
@@ -13,4 +13,4 @@ ctypedef stdint.uint8_t uint8_t
 #ctypedef string.bool bool
 
 cdef extern from "discrete_wavelet_transform.h":
-    void discrete_wavelet_transform_2D(int32_t* data, size_t data_width, size_t data_height, uint8_t transform_levels, uint8_t inverse_flag)
+    void discrete_wavelet_transform_2D(int16_t* data, size_t data_width, size_t data_height, uint8_t transform_levels, uint8_t inverse_flag)
